+++ packages/react-composites/src/composites/CallComposite/components/MediaGallery.tsx
@@ -1,12 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import React, { useMemo } from 'react';
+import React, { useEffect, useMemo, useState } from 'react';
 import { VideoGallery, VideoStreamOptions, OnRenderAvatarCallback } from '@internal/react-components';
 import { usePropsFor } from '../hooks/usePropsFor';
 import { AvatarPersona, AvatarPersonaDataCallback } from '../../common/AvatarPersona';
 import { mergeStyles, Stack } from '@fluentui/react';
-import { useLocalVideoStartTrigger } from '../hooks/useLocalVideoStartTrigger';
+import { getIsPreviewCameraOn } from '../selectors/baseSelectors';
+import { useHandlers } from '../hooks/useHandlers';
+import { useSelector } from '../hooks/useSelector';
 
 const VideoGalleryStyles = {
   root: {
@@ -62,4 +64,31 @@ export const MediaGallery = (props: MediaGalleryProps): JSX.Element => {
   }, [props.onFetchAvatarPersonaData, videoGalleryProps]);
 
   return VideoGalleryMemoized;
+};
+
+/**
+ * @private
+ *
+ * `shouldTransition` is an extra predicate that controls whether this hooks actually transitions the call.
+ * The rule of hooks disallows calling the hook conditionally, so this predicate can be used to make the decision.
+ */
+export const useLocalVideoStartTrigger = (isLocalVideoAvailable: boolean, shouldTransition?: boolean): void => {
+  // Once a call is joined, we need to transition the local preview camera setting into the call.
+  // This logic is needed on any screen that we might join a call from:
+  // - The Media gallery
+  // - The lobby page
+  // - The networkReconnect interstitial that may show at the start of a call.
+  //
+  // @TODO: Can we simply have the callHandlers handle this transition logic.
+  const [isButtonStatusSynced, setIsButtonStatusSynced] = useState(false);
+  const isPreviewCameraOn = useSelector(getIsPreviewCameraOn);
+  const mediaGalleryHandlers = useHandlers(MediaGallery);
+  useEffect(() => {
+    if (shouldTransition !== false) {
+      if (isPreviewCameraOn && !isLocalVideoAvailable && !isButtonStatusSynced) {
+        mediaGalleryHandlers.onStartLocalVideo();
+      }
+      setIsButtonStatusSynced(true);
+    }
+  }, [shouldTransition, isButtonStatusSynced, isPreviewCameraOn, isLocalVideoAvailable, mediaGalleryHandlers]);
 };
